New combobox component (#12973)
## Description 

Generic Implementation of a combobox using
[cmdk](https://cmdk.paco.me/), to allow us to get off of headlessUI
without issue. This is more custom than I'd really like, and if Radix
ever implemented a combobox I'd happily switch to use that instead.

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes