[macros] Update `with_checked_arithmetic` macro, and convert usage of `checked_arithmetic` macro to attribute macro (#13107)
## Description 

Usage of the `checked_arithmetic` macro was causing a number of
quality-of-life issues for development around formatting and LSP
support. However, converting to an attribute-based procedural macro for
adding checked arithmetic operations solves all of these issues (LSPs
and autoformatters should now work with it).

Note that this only works on explicitly declared modules due to
(current) Rust restrictions, and you cannot annotate a file-based module
with this attribute macro.

This also required expanding the applicability of the
`with_checked_arithmetic` macro to work over more forms than it
previously handled, however the change is pretty straightforward.

The one downside to it is that this introduces some slight boilerplate
of

```rust
pub use checked::*;
mod checked {
// what was previously inside the `checked_aritmetic` macro + imports
}
```

## Test Plan 

Added an additional test to make sure the behavior of code within a
module annotated with `with_checked_arithmetic` behaves in the same way
as code within a `checked_arithmetic` block.

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes