[wallet-ext] - assets page fixes (#13125)
## Description 

Fixes a few issues that I noticed with the assets page / kiosks that
didn't present themselves in local development for some reason (maybe
due to minifying / webpack config 🤨)

- While debugging, I noticed that `useOnScreen` was triggering a large
amount of re-renders, so I cleaned it up a bit and it seems to be happy
now :)
- Removed sorting by `display` as this was causing inconsistencies in
rendering between the Asset page and Kiosk detail page. This wasn't
really needed in the first place
- Added a loading spinner to the Kiosk detail page

## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes