[wallet-ext] - asset filtering / more kiosk polish (#13167)
## Description 

Includes various changes to the Assets / NFTs page, mainly: 
- Adds ability to filter by Visual / Non-Visual assets. (note: I split
this out into two pages for now but could probably be rolled back into
one since it's really just a grid / list view)
- Moved the show / hide asset logic into a provider which cleans things
up a bit. We could also potentially expand this in the future if we add
more types of filtering / functionality
- Addressed some remaining feedback / UI polish for kiosks



https://github.com/MystenLabs/sui/assets/122397493/caebcb96-27d2-4a96-a280-98838ae878fc




## Test Plan 

How did you test the new or updated feature?

---
If your changes are not user-facing and not a breaking change, you can
skip the following section. Otherwise, please indicate what changed, and
then add to the Release Notes section as highlighted during the release
process.

### Type of Change (Check all that apply)

- [ ] protocol change
- [ ] user-visible impact
- [ ] breaking change for a client SDKs
- [ ] breaking change for FNs (FN binary must upgrade)
- [ ] breaking change for validators or node operators (must upgrade
binaries)
- [ ] breaking change for on-chain data layout
- [ ] necessitate either a data wipe or data migration

### Release notes